Refactor: replace filter screen with bottom sheet and update UI
This commit replaces the dedicated `FilterScreen` with a `FilterBottomSheet` for a more integrated user experience within the `InsightsScreen`. The elevation of `FilterCategory` has also been increased.

**Key Changes:**

- **Filter Screen to Bottom Sheet:**
    - Removed `FilterScreen.kt`.
    - Introduced `FilterBottomSheet.kt` which implements the filter functionality as a modal bottom sheet.
    - `InsightsScreen` now manages the visibility of `FilterBottomSheet` using `rememberModalBottomSheetState` and a `showBottomSheet` state variable.
    - The Floating Action Button (FAB) on `InsightsScreen` now triggers the display of the `FilterBottomSheet`.
    - The back arrow navigation in the `FilterScreen`'s `TopAppBar` is replaced by the bottom sheet's dismiss functionality and an "Apply" button within the sheet.
- **UI Updates:**
    - **`FilterBottomSheet`:**
        - The title "Filter" is now displayed as a `Text` composable with `MaterialTheme.typography.titleLarge` at the top of the bottom sheet content.
- **State Management:**
    - `InsightsScreen` now collects and passes `filterUiState` and `gameResultFilter` to the `FilterBottomSheet`.
    - The `onApplyClick` lambda in `FilterBottomSheet` is used to hide the sheet.
- **Preview Updates:**
    - Updated `FilterScreenPreview` in the new `FilterBottomSheet.kt` to reflect the new component.
    - Added `filterUiState` and `gameResultFilter` to the `InsightsScreenPreview`.
